Fix for LooselyCoupled template - Removed clearAccessor call from fetchLooselyCoupledRelation
Doctrine_Record stores customAccessors in a static property, not per instance
Doctrine_Templates are stored in the table, not the record, and so are only instantiated (and so setUp etc methods called) once.

Thus, the custom accessor is removed from all present and future instances of the relevant object class by this call, and will never be re-added.
I assume the intention was to remove from the current instance only?